Host model updated i.e communication with guest models + moved Myserver to Model layer and created ModelHandler
