+++ src/main/java/com/example/Scrabble/Model/GameController.java
@@ -1,0 +1,12 @@
+package com.example.Scrabble.Model;
+
+import com.example.Scrabble.ScrabbleServer.MyServer;
+
+class GameController {
+
+    MyServer HostServer;
+
+
+
+
+}
