+++ src/main/java/com/example/Scrabble/Model/HostModel.java
@@ -1,20 +1,19 @@
 package com.example.Scrabble.Model;
 
-import com.example.Scrabble.Server.ScrabbleServer.BookScrabbleHandler;
-import com.example.Scrabble.Server.ScrabbleServer.ClientHandler;
-import com.example.Scrabble.Server.ScrabbleServer.MyServer;
+import com.example.Scrabble.ScrabbleServer.BookScrabbleHandler;
+import com.example.Scrabble.ScrabbleServer.ClientHandler;
+import com.example.Scrabble.ScrabbleServer.MyServer;
 
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.net.ServerSocket;
 import java.net.Socket;
-import java.net.SocketTimeoutException;
 import java.util.Random;
 import java.util.Scanner;
 
 public class HostModel {
     MyServer gameServer;
     String Hostname;
+    MyServer hostServer;
     
     ClientHandler ch;
     boolean stop;
@@ -26,7 +25,7 @@ public class HostModel {
         // create a server for the game
         Random r = new Random();
         gameServer = new MyServer(6000 + r.nextInt(1000), new BookScrabbleHandler());
-
+        hostServer = new MyServer(6000 + r.nextInt(1000), new ModelHandler());
         // host new server for guests
         // this.stop = false;
         // try {
@@ -50,10 +49,12 @@ public class HostModel {
 
     public void connectServer() {
         gameServer.start(); // runs in the background
+        hostServer.start();
     }
 
     public void closeServer() {
         gameServer.close();
+        hostServer.close();
     }
 
     public void handleClient(Socket client) {
@@ -70,10 +71,28 @@ public class HostModel {
         }
 
     }
+    public void test(String query) {
+        try {
+            System.out.println("Connecting to port " + this.hostServer.getPort());
+            Socket server = new Socket("localhost", this.hostServer.getPort());
+            PrintWriter out = new PrintWriter(server.getOutputStream());
+            Scanner in = new Scanner(server.getInputStream());
+            out.println(query);
+            out.flush();
+            String res = in.next();
+            System.out.println(Hostname + " your results are: " + res);
+            in.close();
+            out.close();
+            server.close();
+
+        } catch (IOException e) {
+            System.out.println("IOException in runGame");
+        }
+    }
 
     public void BookActions(String query) {
         try {
-            System.out.println("Connecting to localhost on port " + this.gameServer.getPort());
+            System.out.println("Connecting to port " + this.gameServer.getPort());
             Socket server = new Socket("localhost", this.gameServer.getPort());
             PrintWriter out = new PrintWriter(server.getOutputStream());
             Scanner in = new Scanner(server.getInputStream());
