+++ src/main/java/com/example/Scrabble/Model/HostPlayerModel.java
@@ -1,0 +1,9 @@
+package com.example.Scrabble.Model;
+
+public class HostPlayerModel extends PlayerModel{
+
+    public HostPlayerModel(String playerName, int playerID) {
+        super(playerName, playerID);
+    }
+
+}
