+++ src/main/java/com/example/Scrabble/ScrabbleServer/ClientHandler.java
@@ -1,0 +1,9 @@
+package com.example.Scrabble.ScrabbleServer;
+
+import java.io.InputStream;
+import java.io.OutputStream;
+
+public interface ClientHandler {
+	void handleClient(InputStream inFromclient, OutputStream outToClient);
+	void close();
+}
