--- src/main/java/com/example/Scrabble/Server/ScrabbleServer/ClientHandler.java
@@ -1,9 +1,0 @@
-package com.example.Scrabble.Server.ScrabbleServer;
-
-import java.io.InputStream;
-import java.io.OutputStream;
-
-public interface ClientHandler {
-	void handleClient(InputStream inFromclient, OutputStream outToClient);
-	void close();
-}
