--- src/main/java/com/example/Scrabble/Server/ScrabbleServer/MyServer.java
@@ -1,62 +1,0 @@
-package com.example.Scrabble.Server.ScrabbleServer;
-
-
-import java.io.IOException;
-import java.net.ServerSocket;
-import java.net.Socket;
-import java.net.SocketTimeoutException;
-
-public class MyServer {
-
-    int port;
-    boolean stop;
-    ClientHandler ch;
-
-    public MyServer(int port, ClientHandler ch){
-        this.port = port;
-        this.ch = ch;
-    }
-
-    public void start(){
-        stop = false;
-        new Thread(()->startServer()).start();
-    }
-
-    private void startServer() {
-        try {
-            ServerSocket server = new ServerSocket(port);
-            server.setSoTimeout(1000);
-            while(!stop) {
-                try{
-                    Socket client = server.accept();
-                    ch.handleClient(client.getInputStream(), client.getOutputStream());
-                    ch.close();
-                    client.close();
-                }catch(SocketTimeoutException e) {}
-            }
-            server.close();
-        }catch (IOException e){
-            e.printStackTrace();
-        }
-    }
-    public void close(){
-        stop = true;
-    }
-    public int getPort() {
-        return port;
-    }
-
-    public ClientHandler getCh() {
-        return ch;
-    }
-
-    public boolean isStop() {
-        return stop;
-    }
-
-    public void setStop(boolean stop) {
-        this.stop = stop;
-    }
-
-
-}
