+++ src/main/java/com/example/Scrabble/Test/TestMain.java
@@ -13,6 +13,8 @@ public class TestMain {
         // Random r=new Random();
         // int port=6000+r.nextInt(1000);
         h.connectServer();
+
+        h.test("waka");
         h.BookActions("Q,s1.txt,s2.txt,19393");
